backendapi/requirements.txt
astroid==2.3.2
colorama==0.4.1
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
